requirements.txt
flask_sqlalchemy
flask-cors
gunicorn
psycopg2
pytz
matplotlib
numpy
requests
bs4
lxml
